feature/bank-account-logic - make statementPrinter interface, create ConsoleStatementPrinter realization
